requirements.txt
fastapi==0.99.0
uvicorn==0.23.0
pydantic>=1.7.4,<2.0.0
python-dotenv
replicate
